Move model-view classes from library to library::editor
