+++ libs/librepcb/libraryeditor/pkg/packagepadlistmodel.cpp
@@ -20,13 +20,12 @@
 /*******************************************************************************
  *  Includes
  ******************************************************************************/
-#include "componentsignallistmodel.h"
-
-#include "cmd/cmdcomponentsignaledit.h"
+#include "packagepadlistmodel.h"
 
 #include <librepcb/common/toolbox.h>
 #include <librepcb/common/undocommandgroup.h>
 #include <librepcb/common/undostack.h>
+#include <librepcb/library/pkg/cmd/cmdpackagepadedit.h>
 
 #include <QtCore>
 
@@ -35,46 +34,44 @@
  ******************************************************************************/
 namespace librepcb {
 namespace library {
+namespace editor {
 
 /*******************************************************************************
  *  Constructors / Destructor
  ******************************************************************************/
 
-ComponentSignalListModel::ComponentSignalListModel(QObject* parent) noexcept
+PackagePadListModel::PackagePadListModel(QObject* parent) noexcept
   : QAbstractTableModel(parent),
-    mSignalList(nullptr),
+    mPadList(nullptr),
     mUndoStack(nullptr),
     mNewName(),
-    mNewIsRequired(false),
-    mNewForcedNetName(),
-    mOnEditedSlot(*this, &ComponentSignalListModel::signalListEdited) {
+    mOnEditedSlot(*this, &PackagePadListModel::padListEdited) {
 }
 
-ComponentSignalListModel::~ComponentSignalListModel() noexcept {
+PackagePadListModel::~PackagePadListModel() noexcept {
 }
 
 /*******************************************************************************
  *  Setters
  ******************************************************************************/
 
-void ComponentSignalListModel::setSignalList(
-    ComponentSignalList* list) noexcept {
+void PackagePadListModel::setPadList(PackagePadList* list) noexcept {
   emit beginResetModel();
 
-  if (mSignalList) {
-    mSignalList->onEdited.detach(mOnEditedSlot);
+  if (mPadList) {
+    mPadList->onEdited.detach(mOnEditedSlot);
   }
 
-  mSignalList = list;
+  mPadList = list;
 
-  if (mSignalList) {
-    mSignalList->onEdited.attach(mOnEditedSlot);
+  if (mPadList) {
+    mPadList->onEdited.attach(mOnEditedSlot);
   }
 
   emit endResetModel();
 }
 
-void ComponentSignalListModel::setUndoStack(UndoStack* stack) noexcept {
+void PackagePadListModel::setUndoStack(UndoStack* stack) noexcept {
   mUndoStack = stack;
 }
 
@@ -82,40 +79,40 @@ void ComponentSignalListModel::setUndoStack(UndoStack* stack) noexcept {
  *  Slots
  ******************************************************************************/
 
-void ComponentSignalListModel::addSignal(const QVariant& editData) noexcept {
+void PackagePadListModel::addPad(const QVariant& editData) noexcept {
   Q_UNUSED(editData);
-  if (!mSignalList) {
+  if (!mPadList) {
     return;
   }
 
   try {
     QScopedPointer<UndoCommandGroup> cmd(
-        new UndoCommandGroup(tr("Add component signal(s)")));
+        new UndoCommandGroup(tr("Add package pad(s)")));
+    // if no name is set we search for the next free numerical pad name
+    if (mNewName.isEmpty()) {
+      mNewName = getNextPadNameProposal();
+    }
     foreach (const QString& name, Toolbox::expandRangesInString(mNewName)) {
-      std::shared_ptr<ComponentSignal> sig = std::make_shared<ComponentSignal>(
-          Uuid::createRandom(), validateNameOrThrow(name),
-          SignalRole::passive(), mNewForcedNetName, mNewIsRequired, false,
-          false);
-      cmd->appendChild(new CmdComponentSignalInsert(*mSignalList, sig));
+      std::shared_ptr<PackagePad> pad = std::make_shared<PackagePad>(
+          Uuid::createRandom(), validateNameOrThrow(name));
+      cmd->appendChild(new CmdPackagePadInsert(*mPadList, pad));
     }
     execCmd(cmd.take());
-    mNewName          = QString();
-    mNewIsRequired    = false;
-    mNewForcedNetName = QString();
+    mNewName = QString();
   } catch (const Exception& e) {
     QMessageBox::critical(0, tr("Error"), e.getMsg());
   }
 }
 
-void ComponentSignalListModel::removeSignal(const QVariant& editData) noexcept {
-  if (!mSignalList) {
+void PackagePadListModel::removePad(const QVariant& editData) noexcept {
+  if (!mPadList) {
     return;
   }
 
   try {
-    Uuid uuid = Uuid::fromString(editData.toString());
-    std::shared_ptr<ComponentSignal> sig = mSignalList->get(uuid);
-    execCmd(new CmdComponentSignalRemove(*mSignalList, sig.get()));
+    Uuid                        uuid = Uuid::fromString(editData.toString());
+    std::shared_ptr<PackagePad> pad  = mPadList->get(uuid);
+    execCmd(new CmdPackagePadRemove(*mPadList, pad.get()));
   } catch (const Exception& e) {
     QMessageBox::critical(0, tr("Error"), e.getMsg());
   }
@@ -125,33 +122,32 @@ void ComponentSignalListModel::removeSignal(const QVariant& editData) noexcept {
  *  Inherited from QAbstractItemModel
  ******************************************************************************/
 
-int ComponentSignalListModel::rowCount(const QModelIndex& parent) const {
-  if (!parent.isValid() && mSignalList) {
-    return mSignalList->count() + 1;
+int PackagePadListModel::rowCount(const QModelIndex& parent) const {
+  if (!parent.isValid() && mPadList) {
+    return mPadList->count() + 1;
   }
   return 0;
 }
 
-int ComponentSignalListModel::columnCount(const QModelIndex& parent) const {
+int PackagePadListModel::columnCount(const QModelIndex& parent) const {
   if (!parent.isValid()) {
     return _COLUMN_COUNT;
   }
   return 0;
 }
 
-QVariant ComponentSignalListModel::data(const QModelIndex& index,
-                                        int                role) const {
-  if (!index.isValid() || !mSignalList) {
+QVariant PackagePadListModel::data(const QModelIndex& index, int role) const {
+  if (!index.isValid() || !mPadList) {
     return QVariant();
   }
 
-  std::shared_ptr<ComponentSignal> item = mSignalList->value(index.row());
+  std::shared_ptr<PackagePad> item = mPadList->value(index.row());
   switch (index.column()) {
     case COLUMN_NAME: {
       QString name     = item ? *item->getName() : mNewName;
       bool    showHint = (!item) && mNewName.isEmpty();
       QString hint =
-          tr("Signal name (may contain ranges like \"%1\")").arg("1..5");
+          tr("Pad name (may contain ranges like \"%1\")").arg("1..5");
       switch (role) {
         case Qt::DisplayRole:
           return showHint ? hint : name;
@@ -171,31 +167,6 @@ QVariant ComponentSignalListModel::data(const QModelIndex& index,
           return QVariant();
       }
     }
-    case COLUMN_ISREQUIRED: {
-      bool required = item ? item->isRequired() : mNewIsRequired;
-      switch (role) {
-        case Qt::DisplayRole:
-          return required ? tr("Required") : tr("Optional");
-        case Qt::CheckStateRole:
-          return required ? Qt::Checked : Qt::Unchecked;
-        case Qt::ToolTipRole:
-          return required ? tr("Leaving this signal unconnected in schematics "
-                               "produces an ERC error.")
-                          : tr("Leaving this signal unconnected in schematics "
-                               "is allowed.");
-        default:
-          return QVariant();
-      }
-    }
-    case COLUMN_FORCEDNETNAME: {
-      switch (role) {
-        case Qt::DisplayRole:
-        case Qt::EditRole:
-          return item ? item->getForcedNetName() : mNewForcedNetName;
-        default:
-          return QVariant();
-      }
-    }
     case COLUMN_ACTIONS: {
       switch (role) {
         case Qt::EditRole:
@@ -211,29 +182,25 @@ QVariant ComponentSignalListModel::data(const QModelIndex& index,
   return QVariant();
 }
 
-QVariant ComponentSignalListModel::headerData(int             section,
-                                              Qt::Orientation orientation,
-                                              int             role) const {
+QVariant PackagePadListModel::headerData(int             section,
+                                         Qt::Orientation orientation,
+                                         int             role) const {
   if (orientation == Qt::Horizontal) {
     if (role == Qt::DisplayRole) {
       switch (section) {
         case COLUMN_NAME:
           return tr("Name");
-        case COLUMN_ISREQUIRED:
-          return tr("Connection");
-        case COLUMN_FORCEDNETNAME:
-          return tr("Forced Net");
         default:
           return QVariant();
       }
     }
   } else if (orientation == Qt::Vertical) {
-    if (mSignalList && (role == Qt::DisplayRole)) {
-      std::shared_ptr<ComponentSignal> item = mSignalList->value(section);
+    if (mPadList && (role == Qt::DisplayRole)) {
+      std::shared_ptr<PackagePad> item = mPadList->value(section);
       return item ? item->getUuid().toStr().left(8) : tr("New:");
-    } else if (mSignalList && (role == Qt::ToolTipRole)) {
-      std::shared_ptr<ComponentSignal> item = mSignalList->value(section);
-      return item ? item->getUuid().toStr() : tr("Add a new signal");
+    } else if (mPadList && (role == Qt::ToolTipRole)) {
+      std::shared_ptr<PackagePad> item = mPadList->value(section);
+      return item ? item->getUuid().toStr() : tr("Add a new pad");
     } else if (role == Qt::TextAlignmentRole) {
       return QVariant(Qt::AlignRight | Qt::AlignVCenter);
     } else if (role == Qt::FontRole) {
@@ -247,27 +214,25 @@ QVariant ComponentSignalListModel::headerData(int             section,
   return QVariant();
 }
 
-Qt::ItemFlags ComponentSignalListModel::flags(const QModelIndex& index) const {
+Qt::ItemFlags PackagePadListModel::flags(const QModelIndex& index) const {
   Qt::ItemFlags f = QAbstractTableModel::flags(index);
-  if (index.isValid() && (index.column() == COLUMN_ISREQUIRED)) {
-    f |= Qt::ItemIsUserCheckable;
-  } else if (index.isValid() && (index.column() != COLUMN_ACTIONS)) {
+  if (index.isValid()) {
     f |= Qt::ItemIsEditable;
   }
   return f;
 }
 
-bool ComponentSignalListModel::setData(const QModelIndex& index,
-                                       const QVariant& value, int role) {
-  if (!mSignalList) {
+bool PackagePadListModel::setData(const QModelIndex& index,
+                                  const QVariant& value, int role) {
+  if (!mPadList) {
     return false;
   }
 
   try {
-    std::shared_ptr<ComponentSignal> item = mSignalList->value(index.row());
-    QScopedPointer<CmdComponentSignalEdit> cmd;
+    std::shared_ptr<PackagePad>       item = mPadList->value(index.row());
+    QScopedPointer<CmdPackagePadEdit> cmd;
     if (item) {
-      cmd.reset(new CmdComponentSignalEdit(*item));
+      cmd.reset(new CmdPackagePadEdit(*item));
     }
     if ((index.column() == COLUMN_NAME) && role == Qt::EditRole) {
       QString name        = value.toString().trimmed();
@@ -283,22 +248,6 @@ bool ComponentSignalListModel::setData(const QModelIndex& index,
         } else {
           mNewName = name;  // contains ranges -> keep them!
         }
-      }
-    } else if ((index.column() == COLUMN_ISREQUIRED) &&
-               role == Qt::CheckStateRole) {
-      bool required = value.toInt() == Qt::Checked;
-      if (cmd) {
-        cmd->setIsRequired(required);
-      } else {
-        mNewIsRequired = required;
-      }
-    } else if ((index.column() == COLUMN_FORCEDNETNAME) &&
-               role == Qt::EditRole) {
-      QString forcedNetName = cleanForcedNetName(value.toString());
-      if (cmd) {
-        cmd->setForcedNetName(forcedNetName);
-      } else {
-        mNewForcedNetName = forcedNetName;
       }
     } else {
       return false;  // do not execute command!
@@ -319,32 +268,31 @@ bool ComponentSignalListModel::setData(const QModelIndex& index,
  *  Private Methods
  ******************************************************************************/
 
-void ComponentSignalListModel::signalListEdited(
-    const ComponentSignalList& list, int index,
-    const std::shared_ptr<const ComponentSignal>& signal,
-    ComponentSignalList::Event                    event) noexcept {
+void PackagePadListModel::padListEdited(
+    const PackagePadList& list, int index,
+    const std::shared_ptr<const PackagePad>& pad,
+    PackagePadList::Event                    event) noexcept {
   Q_UNUSED(list);
-  Q_UNUSED(signal);
+  Q_UNUSED(pad);
   switch (event) {
-    case ComponentSignalList::Event::ElementAdded:
+    case PackagePadList::Event::ElementAdded:
       beginInsertRows(QModelIndex(), index, index);
       endInsertRows();
       break;
-    case ComponentSignalList::Event::ElementRemoved:
+    case PackagePadList::Event::ElementRemoved:
       beginRemoveRows(QModelIndex(), index, index);
       endRemoveRows();
       break;
-    case ComponentSignalList::Event::ElementEdited:
+    case PackagePadList::Event::ElementEdited:
       dataChanged(this->index(index, 0), this->index(index, _COLUMN_COUNT - 1));
       break;
     default:
-      qWarning() << "Unhandled switch-case in "
-                    "ComponentSignalListModel::signalListEdited()";
+      qWarning() << "Unhandled switch-case in PackagePadList::padListEdited()";
       break;
   }
 }
 
-void ComponentSignalListModel::execCmd(UndoCommand* cmd) {
+void PackagePadListModel::execCmd(UndoCommand* cmd) {
   if (mUndoStack) {
     mUndoStack->execCmd(cmd);
   } else {
@@ -353,30 +301,28 @@ void ComponentSignalListModel::execCmd(UndoCommand* cmd) {
   }
 }
 
-CircuitIdentifier ComponentSignalListModel::validateNameOrThrow(
+CircuitIdentifier PackagePadListModel::validateNameOrThrow(
     const QString& name) const {
-  if (mSignalList && mSignalList->contains(name)) {
+  if (mPadList && mPadList->contains(name)) {
     throw RuntimeError(
         __FILE__, __LINE__,
-        QString(tr("There is already a signal with the name \"%1\"."))
-            .arg(name));
+        QString(tr("There is already a pad with the name \"%1\".")).arg(name));
   }
   return CircuitIdentifier(name);  // can throw
 }
 
-QString ComponentSignalListModel::cleanForcedNetName(
-    const QString& name) noexcept {
-  // Same as cleanCircuitIdentifier(), but allowing '{' and '}' because it's
-  // allowed to have attribute placeholders in a forced net name. Also remove
-  // spaces because they must not be replaced by underscores inside {{ and }}.
-  return Toolbox::cleanUserInputString(
-      name, QRegularExpression("[^-a-zA-Z0-9_+/!?@#$\\{\\}]"), true, false,
-      false, "");
+QString PackagePadListModel::getNextPadNameProposal() const noexcept {
+  int i = 1;
+  while (mPadList->contains(QString::number(i))) {
+    ++i;
+  }
+  return QString::number(i);
 }
 
 /*******************************************************************************
  *  End of File
  ******************************************************************************/
 
+}  // namespace editor
 }  // namespace library
 }  // namespace librepcb
